backend/app/requirements_deploy.txt
fastapi
uvicorn
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
alembic
